# docs/openapi-dify.yaml
openapi: 3.0.3
info:
  title: Outdoor SaaS API
  version: 1.0.0
  description: |
    户外广告管理系统 REST API
    用于 Dify AI 助手集成

    ## 鉴权说明
    本 API 支持 API Key 鉴权方式，用于 Dify 自定义工具等外部系统调用。

    ### 获取 API Key
    请联系系统管理员获取 API Key。

    ### 使用方式
    在请求头中添加：`X-API-Key: your-api-key`

    ### Dify 配置
    在 Dify 中创建自定义工具时：
    1. 鉴权方式选择 "API Key"
    2. 鉴权字段填写 `X-API-Key`
    3. API Key 值填写管理员提供的密钥
servers:
  - url: http://192.168.31.34:16000
    description: 本地开发环境（frp内网穿透）
paths:
...
  /api/frame/{id}:
    get:
      tags:
        - 框架管理
      summary: 根据ID查询框架
      description: 获取指定ID的框架详细信息
      operationId: get_frame
      parameters:
        - name: id
          in: path
          required: true
          description: 框架ID
          schema:
            type: integer
      responses:
        '200':
          description: 成功

    delete:
      tags:
        - 框架管理
      summary: 删除框架
      description: 根据ID删除指定框架
      operationId: delete_frame
      parameters:
        - name: id
          in: path
          required: true
          description: 框架ID
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
...
  /api/frame/filter/page:
    post:
      tags:
        - 框架管理
      summary: 分页查询框架
      description: 根据条件分页查询框架列表，支持多种查询条件组合
      operationId: filter_frames
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FrameQueryParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultPageResultFrame'

  # ========== 方案管理 ==========
...
  /api/plan/{id}:
    get:
      tags:
        - 方案管理
      summary: 根据ID查询方案
      description: 获取指定ID的方案详细信息
      operationId: get_plan
      parameters:
        - name: id
          in: path
          required: true
          description: 方案ID
          schema:
            type: integer
      responses:
        '200':
          description: 成功

    delete:
      tags:
        - 方案管理
      summary: 删除方案
      description: 根据ID删除指定方案
      operationId: delete_plan
      parameters:
        - name: id
          in: path
          required: true
          description: 方案ID
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
...
  /api/plan/filter/page:
    post:
      tags:
        - 方案管理
      summary: 分页查询方案
      description: 根据条件分页查询方案列表，支持多种查询条件组合
      operationId: filter_plans
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanQueryParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultPageResultPlan'

  # ========== 方案关联管理 ==========
...
  /api/plan-community/{id}:
    delete:
      tags:
        - 方案关联管理
      summary: 删除方案社区关联
      description: 从方案中移除社区
      operationId: delete_plan_community
      parameters:
        - name: id
          in: path
          required: true
          description: 关联ID
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultInteger'
...
  /api/plan-frame/filter/page:
    post:
      tags:
        - 方案关联管理
      summary: 分页查询方案框架关联
      description: 分页查询方案与框架的关联关系，支持多种查询条件组合
      operationId: filter_plan_frames
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanFrameQueryParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultPageResultPlanFrame'
  /api/plan-frame:
    post:
      tags:
        - 方案关联管理
      summary: 创建方案框架关联
      description: 将框架添加到方案
      operationId: create_plan_frame
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanFrame'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultInteger'
  /api/plan-frame/batch:
    post:
      tags:
        - 方案关联管理
      summary: 批量创建方案框架关联
      description: 批量将多个框架添加到方案，支持统一设置投放日期和方案社区关联
      operationId: batch_create_plan_frame
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanFrameBatchCreateDTO'
      responses:
        '200':
          description: 批量创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultInteger'
  /api/plan-frame/{id}:
    delete:
      tags:
        - 方案关联管理
      summary: 删除方案框架关联
      description: 从方案中移除框架
      operationId: delete_plan_frame
      parameters:
        - name: id
          in: path
          required: true
          description: 关联ID
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultInteger'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 请输入 JWT Token，格式：Bearer {token}
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key 鉴权，用于 Dify 等外部工具调用。请联系管理员获取 API Key。
  schemas:
...
    ResultInteger:
      type: object
      properties:
        code:
          type: integer
          description: 响应码，200表示成功
        message:
          type: string
          description: 响应消息
        data:
          type: integer
          description: 返回的整数数据（通常为ID或影响行数）
        timestamp:
          type: integer
          format: int64
          description: 时间戳
...
    PageRequest:
      type: object
      description: 分页请求基类，所有分页查询参数都继承此类
      properties:
        pageNum:
          type: integer
          description: 当前页码，从1开始，默认为1
          minimum: 1
          default: 1
        pageSize:
          type: integer
          description: 每页大小，默认为10，最大1000
          minimum: 1
          maximum: 1000
          default: 10
        orderBy:
          type: string
          description: 排序字段名称
        order:
          type: string
          description: 排序方式，asc为升序，desc为降序，默认为desc
          enum: [asc, desc]
          default: desc

    # ========== 基础查询参数 ==========
    BaseQueryParam:
      type: object
      description: 查询参数基类，包含分页、时间范围、关键词搜索等通用参数
      allOf:
        - $ref: '#/components/schemas/PageRequest'
      properties:
        keyword:
          type: string
          description: 关键词搜索，用于名称、编号等字段的模糊搜索
        startTime:
          type: string
          format: date-time
          description: 开始时间，用于时间范围查询
        endTime:
          type: string
          format: date-time
          description: 结束时间，用于时间范围查询
        createdAtBegin:
          type: string
          format: date-time
          description: 创建时间开始，用于查询创建时间在此时间之后的数据
        createdAtEnd:
          type: string
          format: date-time
          description: 创建时间结束，用于查询创建时间在此时间之前的数据
        updatedAtBegin:
          type: string
          format: date-time
          description: 更新时间开始，用于查询更新时间在此时间之后的数据
        updatedAtEnd:
          type: string
          format: date-time
          description: 更新时间结束，用于查询更新时间在此时间之前的数据
        statusList:
          type: string
          description: 状态列表，用于IN查询，多个状态用逗号分隔，如"1,2,3"
        idList:
          type: string
          description: ID列表，用于IN查询，多个ID用逗号分隔，如"1,2,3"

    # ========== 社区 ==========
...
    Frame:
      type: object
      properties:
        id:
          type: integer
          description: 框架ID
        frameNo:
          type: string
          description: 框架编号
        building:
          type: string
          description: 楼栋
        unit:
          type: string
          description: 单元
        elevator:
          type: string
          description: 电梯
        innerPosition:
          type: integer
          description: 内外位置
        communityId:
          type: integer
          description: 所属社区ID
...
    FrameQueryParam:
      type: object
      description: 框架媒体分页查询参数，支持精确匹配、模糊搜索、IN查询等多种查询方式，支持关联社区表查询
      allOf:
        - $ref: '#/components/schemas/BaseQueryParam'
      properties:
        frameNo:
          type: string
          description: 框架编号（精确匹配）
        frameNoLike:
          type: string
          description: 框架编号模糊搜索，支持部分匹配
        communityId:
          type: integer
          description: 社区ID（精确匹配）
        communityIdList:
          type: array
          description: 社区ID列表（IN查询），可同时查询多个社区
          items:
            type: integer
        building:
          type: string
          description: 楼栋（精确匹配）
        buildingLike:
          type: string
          description: 楼栋模糊搜索，支持部分匹配
        unit:
          type: string
          description: 单元（精确匹配）
        unitLike:
          type: string
          description: 单元模糊搜索，支持部分匹配
        elevator:
          type: string
          description: 电梯（精确匹配）
        elevatorLike:
          type: string
          description: 电梯模糊搜索，支持部分匹配
        innerPosition:
          type: integer
          description: 内外位置（精确匹配）
        innerPositionList:
          type: array
          description: 内外位置列表（IN查询），可同时查询多个位置
          items:
            type: integer
        city:
          type: string
          description: 城市（关联社区表查询），将根据城市筛选所属社区的框架
        buildingNameLike:
          type: string
          description: 楼栋名称模糊搜索（关联社区表），将根据楼栋名称模糊匹配所属社区的框架

    # ========== 方案 ==========
    Plan:
      type: object
      properties:
        id:
          type: integer
          description: 方案ID
        planNo:
          type: string
          description: 方案编号
        planName:
          type: string
          description: 方案名称
        customer:
          type: string
          description: 客户名称
        salesType:
          type: integer
          description: 销售类型
        mediaType:
          type: integer
          description: 媒体类型
        releaseStatus:
          type: integer
          description: 发布状态
        releaseDateBegin:
          type: string
          format: date
          description: 投放开始日期
        releaseDateEnd:
          type: string
          format: date
          description: 投放结束日期
        salesPerson:
          type: string
          description: 销售人员
        budget:
          type: integer
          format: int64
          description: 预算金额（单位：元）
        estimatedReach:
          type: integer
          description: 预计触达人数
...
    PlanQueryParam:
      type: object
      description: 投放方案分页查询参数，支持精确匹配、模糊搜索、IN查询、范围查询等多种查询方式
      allOf:
        - $ref: '#/components/schemas/BaseQueryParam'
      properties:
        planNo:
          type: string
          description: 方案编号（精确匹配）
        planNoLike:
          type: string
          description: 方案编号模糊搜索，支持部分匹配
        planName:
          type: string
          description: 方案名称（精确匹配）
        planNameLike:
          type: string
          description: 方案名称模糊搜索，支持部分匹配
        customer:
          type: string
          description: 客户名称（精确匹配）
        customerLike:
          type: string
          description: 客户名称模糊搜索，支持部分匹配
        releaseStatus:
          type: integer
          description: 投放状态（精确匹配），具体状态码含义请参考枚举定义
        releaseStatusList:
          type: array
          description: 投放状态列表（IN查询），可同时查询多个投放状态
          items:
            type: integer
        salesType:
          type: integer
          description: 销售类型（精确匹配），具体类型码含义请参考枚举定义
        salesTypeList:
          type: array
          description: 销售类型列表（IN查询），可同时查询多个销售类型
          items:
            type: integer
        mediaType:
          type: integer
          description: 媒体类型（精确匹配），具体类型码含义请参考枚举定义
        mediaTypeList:
          type: array
          description: 媒体类型列表（IN查询），可同时查询多个媒体类型
          items:
            type: integer
        releaseDateBeginFrom:
          type: string
          format: date
          description: 投放开始日期范围起点（大于等于），用于筛选投放开始日期在此日期之后的数据
        releaseDateBeginTo:
          type: string
          format: date
          description: 投放开始日期范围终点（小于等于），用于筛选投放开始日期在此日期之前的数据
        releaseDateEndFrom:
          type: string
          format: date
          description: 投放结束日期范围起点（大于等于），用于筛选投放结束日期在此日期之后的数据
        releaseDateEndTo:
          type: string
          format: date
          description: 投放结束日期范围终点（小于等于），用于筛选投放结束日期在此日期之前的数据
        salesPerson:
          type: string
          description: 销售人员（精确匹配）
        salesPersonLike:
          type: string
          description: 销售人员模糊搜索，支持部分匹配
        budgetMin:
          type: integer
          format: int64
          description: 预算最小值（大于等于），用于筛选预算在此金额以上的方案
        budgetMax:
          type: integer
          format: int64
          description: 预算最大值（小于等于），用于筛选预算在此金额以下的方案
        estimatedReachMin:
          type: integer
          description: 预计触达人数最小值（大于等于），用于筛选预计触达人数在此数值以上的方案
        estimatedReachMax:
          type: integer
          description: 预计触达人数最大值（小于等于），用于筛选预计触达人数在此数值以下的方案

    # ========== 分页响应 ==========
...
    ResultPageResultFrame:
      type: object
      description: 框架分页查询响应
      properties:
        code:
          type: integer
          description: 响应码，200表示成功
        message:
          type: string
          description: 响应消息
        data:
          type: object
          description: 分页结果
          properties:
            pageNum:
              type: integer
              description: 当前页码
            pageSize:
              type: integer
              description: 每页大小
            total:
              type: integer
              format: int64
              description: 总记录数
            pages:
              type: integer
              description: 总页数
            list:
              type: array
              description: 框架列表
              items:
                $ref: '#/components/schemas/Frame'
        timestamp:
          type: integer
          format: int64
          description: 时间戳
    ResultPageResultPlan:
      type: object
      description: 方案分页查询响应
      properties:
        code:
          type: integer
          description: 响应码，200表示成功
        message:
          type: string
          description: 响应消息
        data:
          type: object
          description: 分页结果
          properties:
            pageNum:
              type: integer
              description: 当前页码
            pageSize:
              type: integer
              description: 每页大小
            total:
              type: integer
              format: int64
              description: 总记录数
            pages:
              type: integer
              description: 总页数
            list:
              type: array
              description: 方案列表
              items:
                $ref: '#/components/schemas/Plan'
        timestamp:
          type: integer
          format: int64
          description: 时间戳

    # ========== 方案关联 ==========
...
    PlanFrame:
      type: object
      description: 方案框架关联实体
      properties:
        id:
          type: integer
          description: 关联ID
        planId:
          type: integer
          description: 方案ID
        frameId:
          type: integer
          description: 框架ID
        planCommunityId:
          type: integer
          description: 方案社区关联ID，表示该框架属于方案中的哪个社区
        releaseDateBegin:
          type: string
          format: date
          description: 在该框架的投放开始日期
        releaseDateEnd:
          type: string
          format: date
          description: 在该框架的投放结束日期
        releaseStatus:
          type: integer
          description: 发布状态
    PlanFrameBatchCreateDTO:
      type: object
      description: 方案框架关联批量创建参数，用于一次性将多个框架添加到方案
      properties:
        planId:
          type: integer
          description: 方案ID（必填），指定要将框架添加到哪个方案
        frameIds:
          type: array
          description: 框架ID列表（必填），要添加到方案的框架ID列表，至少包含一个
          items:
            type: integer
          minItems: 1
        planCommunityId:
          type: integer
          description: 方案社区关联ID（可选），表示该框架属于方案中的哪个社区
        releaseDateBegin:
          type: string
          format: date
          description: 投放开始日期（可选），默认为空
        releaseDateEnd:
          type: string
          format: date
          description: 投放结束日期（可选），默认为空
        releaseStatus:
          type: integer
          description: 发布状态（可选），默认为空
      required:
        - planId
        - frameIds
    PlanFrameQueryParam:
      type: object
      description: 方案框架关联分页查询参数，支持精确匹配、IN查询、范围查询等多种查询方式
      allOf:
        - $ref: '#/components/schemas/BaseQueryParam'
      properties:
        planId:
          type: integer
          description: 方案ID（精确匹配）
        planIdList:
          type: array
          description: 方案ID列表（IN查询），可同时查询多个方案
          items:
            type: integer
        frameId:
          type: integer
          description: 框架ID（精确匹配）
        frameIdList:
          type: array
          description: 框架ID列表（IN查询），可同时查询多个框架
          items:
            type: integer
        planCommunityId:
          type: integer
          description: 方案社区关联ID（精确匹配），用于查询特定方案社区下的框架
        planCommunityIdList:
          type: array
          description: 方案社区关联ID列表（IN查询），可同时查询多个方案社区
          items:
            type: integer
        releaseStatus:
          type: integer
          description: 发布状态（精确匹配），具体状态码含义请参考枚举定义
        releaseStatusList:
          type: array
          description: 发布状态列表（IN查询），可同时查询多个发布状态
          items:
            type: integer
        releaseDateBeginFrom:
          type: string
          format: date
          description: 发布开始日期范围起点（大于等于），用于筛选发布开始日期在此日期之后的数据
        releaseDateBeginTo:
          type: string
          format: date
          description: 发布开始日期范围终点（小于等于），用于筛选发布开始日期在此日期之前的数据
        releaseDateEndFrom:
          type: string
          format: date
          description: 发布结束日期范围起点（大于等于），用于筛选发布结束日期在此日期之后的数据
        releaseDateEndTo:
          type: string
          format: date
          description: 发布结束日期范围终点（小于等于），用于筛选发布结束日期在此日期之前的数据
    ResultPageResultPlanFrame:
      type: object
      description: 方案框架关联分页查询响应
      properties:
        code:
          type: integer
          description: 响应码，200表示成功
        message:
          type: string
          description: 响应消息
        data:
          type: object
          description: 分页结果
          properties:
            pageNum:
              type: integer
              description: 当前页码
            pageSize:
              type: integer
              description: 每页大小
            total:
              type: integer
              format: int64
              description: 总记录数
            pages:
              type: integer
              description: 总页数
            list:
              type: array
              description: 方案框架关联列表
              items:
                $ref: '#/components/schemas/PlanFrame'
        timestamp:
          type: integer
          format: int64
          description: 时间戳

# 全局安全设置
security:
  - ApiKeyAuth: []